fix tests for latest compiler flags

+++ cosas/test/oscillator.cpp
@@ -18,7 +18,7 @@ TEST_CASE("Oscillator, Detune") {
   AbsDexOsc o1 = AbsDexOsc(4400, w, w.sine_gamma_1);
   RelDexOsc o2 = RelDexOsc(w, w.sine_gamma_1, o1.get_freq_param(), 1, 1);
   CHECK(o2.next(1000, 0) == -32417);
-  o2.get_freq_param().get_det_param().set(0.9);
+  o2.get_freq_param().get_det_param().set(0.9f);
   CHECK(o2.next(1000, 0) != -32417);
 }
 
@@ -28,11 +28,11 @@ TEST_CASE("Oscillator, AbsDexOsc") {
   AbsDexOsc o = AbsDexOsc(440, w, w.sine_gamma_1);
   // there are table_size samples a second.  a 440hz note takes 1/440s
   // for one cycle, which is sample_rate/440 samples.
-  CHECK(o.next(0.0 * SAMPLE_RATE / 440, 0) == 0);
-  CHECK(abs(o.next(0.25 * SAMPLE_RATE / 440, 0) - SAMPLE_MAX) < 1000);
-  CHECK(abs(o.next(0.25 * SAMPLE_RATE / 440, 0) - 0) < 1000);
-  CHECK(abs(o.next(0.25 * SAMPLE_RATE / 440, 0) - SAMPLE_MIN) < 1000);
-  CHECK(abs(o.next(0.25 * SAMPLE_RATE / 440, 0) - 0) < 1000);
+  CHECK(o.next(static_cast<int32_t>(0.0f * SAMPLE_RATE / 440), 0) == 0);
+  CHECK(abs(o.next(static_cast<int32_t>(0.25f * SAMPLE_RATE / 440), 0) - SAMPLE_MAX) < 1000);
+  CHECK(abs(o.next(static_cast<int32_t>(0.25f * SAMPLE_RATE / 440), 0) - 0) < 1000);
+  CHECK(abs(o.next(static_cast<int32_t>(0.25f * SAMPLE_RATE / 440), 0) - SAMPLE_MIN) < 1000);
+  CHECK(abs(o.next(static_cast<int32_t>(0.25f * SAMPLE_RATE / 440), 0) - 0) < 1000);
 }
 
 
@@ -40,9 +40,9 @@ TEST_CASE("Oscillator, AbsPolyOsc") {
   AbsPolyOsc o = AbsPolyOsc(440, PolyTable::SINE, 0, QUARTER_TABLE_SIZE);
   // there are table_size samples a second.  a 440hz note takes 1/440s
   // for one cycle, which is sample_rate/440 samples.
-  CHECK(o.next(0.0 * SAMPLE_RATE / 440, 0) == 0);
-  CHECK(abs(o.next(0.25 * SAMPLE_RATE / 440, 0) - SAMPLE_MAX) < 1000);
-  CHECK(abs(o.next(0.25 * SAMPLE_RATE / 440, 0) - 0) < 1000);
-  CHECK(abs(o.next(0.25 * SAMPLE_RATE / 440, 0) - SAMPLE_MIN) < 1000);
-  CHECK(abs(o.next(0.25 * SAMPLE_RATE / 440, 0) - 0) < 1000);
+  CHECK(o.next(static_cast<int32_t>(0.0f * SAMPLE_RATE / 440), 0) == 0);
+  CHECK(abs(o.next(static_cast<int32_t>(0.25f * SAMPLE_RATE / 440), 0) - SAMPLE_MAX) < 1000);
+  CHECK(abs(o.next(static_cast<int32_t>(0.25f * SAMPLE_RATE / 440), 0) - 0) < 1000);
+  CHECK(abs(o.next(static_cast<int32_t>(0.25f * SAMPLE_RATE / 440), 0) - SAMPLE_MIN) < 1000);
+  CHECK(abs(o.next(static_cast<int32_t>(0.25f * SAMPLE_RATE / 440), 0) - 0) < 1000);
 }
